Added example dir
Added example directory which contains the conf.yaml and exex.yaml file. Attached a MD file to help user understand the execution. Minor tweaks in rexe.py file to run the whole file end-to-end. Tested the code with two servers from local machine and executed a few commands in both the servers.

Signed-off-by: Ayush Ujjwal <aujjwal@aujjwal.remote.csb>